==> 01_extract_introns_from_gencode.ipynb <==
# 01_extract_introns_from_gencode.ipynb

This module extracts all the introns from GENCODE v44 using the Biopython's GFF library

import pandas as pd
from BCBio import GFF

in_file = "./ftp-data/gencode.v44.annotation.gtf"
# examiner = GFFExaminer()

limit_info = {"gff_type": ["transcript", "exon"]}
current_gene = None
last_index = None
introns_lst = []
with open(in_file) as handle:
    for rec in GFF.parse(handle, limit_info=limit_info, target_lines=100):
        for feature in rec.features:
            loop = feature.sub_features if feature.strand == 1 else feature.sub_features[::-1]
            first_exon = True
            for sub_features in loop:
                if sub_features.type == "exon":
                    if not first_exon:
                        intron_dict = {
                            "chr": rec.id,
                            "gene_id": sub_features.qualifiers["gene_id"][0], # this for debugging only
                            "transcript_id": sub_features.qualifiers["transcript_id"][0], # this for debugging only
                            "start": last_index,
                            "end": sub_features.location.start.position,
                            "strand": "+" if feature.strand == 1 else "-",
                            "prev_exon_id": previous_exon_name, # this for debugging only
                            "next_exon_id": sub_features.qualifiers["exon_id"][0] # this for debugging only
                        }
                        introns_lst.append(intron_dict)
                    last_index = sub_features.location.end.position
                    previous_exon_name = sub_features.qualifiers["exon_id"][0]

                    first_exon = False

df_introns = pd.DataFrame(introns_lst.copy())
df_introns

df_introns.drop_duplicates(subset=["chr", "start", "end", "strand"], inplace=True)
df_introns

df_introns = df_introns.drop(columns=["gene_id", "transcript_id", "prev_exon_id", "next_exon_id"])
df_introns

df_introns.to_csv("data/01_gencode_introns.tsv", index=False, sep="\t")

# Experimental code:

df_introns = pd.DataFrame(introns_lst.copy())
df_introns

# df_introns = pd.DataFrame(introns_lst.copy())
df_introns.to_csv("data/gencode_v44_introns.csv", index=False)
df_introns

df_introns_1_based = df_introns.copy()
df_introns_1_based.intron_start = df_introns_1_based.intron_start + 1
df_introns_1_based

df_introns_1_based[df_introns_1_based.intron_start == 962048]

df_introns_1_based[df_introns_1_based.intron_start == 155357520]

df_introns_1_based[df_introns_1_based.intron_start == 155759144]
