==> 02_extract_negative_introns.ipynb <==
# 02_extract_negative_introns.ipynb

This module extracts all the rejected introns that was reviewed manually (some introns are now accepted, which we will not include them as rejected)

import pandas as pd

df_gencode = pd.read_csv("data/01_gencode_introns.tsv", sep="\t")
df_gencode

# https://github.com/kennylam8/GSoC_ML_gene_annot/blob/main/gene_annot_test.tsv
df_manually_reviewed = pd.read_csv("data/resources/manually_reviewed_introns.tsv", sep="\t")
df_manually_reviewed

df_original_test_rejected = df_manually_reviewed[df_manually_reviewed.outcome == "rejected"]

drop_idx_lst = []
append_lst = []
for index, row in df_original_test_rejected.iterrows():
	coords_split = row.coords.split(":")
	chromosome = coords_split[0]
	start = int(coords_split[1].split("-")[0])-1
	end = int(coords_split[1].split("-")[1])
	strand = "+" if str(coords_split[2]) == "1" else "-"
	if df_gencode[(df_gencode.strand == strand) & (df_gencode.start == start) & (df_gencode.end == end) & (df_gencode.chr == chromosome)].empty:
		append_dict = {
			"chr": chromosome,
			"start": start,
			"end": end,
			"strand": strand,
		}
		append_lst.append(append_dict)
	else:
		# print(df_gencode[(df_gencode.strand == strand) & (df_gencode.start == start-1) & (df_gencode.end == end) & (df_gencode.chr == chromosome)])
		drop_idx_lst.append(index)


## These are the entries that were rejected from the original dict but accepted in the latest version

df_original_test_rejected.loc[drop_idx_lst]
# chr10:72275560-72275899:1
# chr11:26567140-26569473:-1
# chr15:61861626-61890833:1
# chr15:61890863-61894121:1
# chr7:129925550-129937707:-1
# chr5:180835003-180835421:1

## We form the new Dataframe and output it to .tsv from the "still rejected" introns

df_output = pd.DataFrame.from_dict(append_lst)
df_output.to_csv("data/02_manually_rejected_introns.tsv", sep="\t", index=False)

len(df_output)
