==> 2 - Combining Accepted and Rejected set.ipynb <==
import pandas as pd

df_accepted = pd.read_csv("data/1-rmv_dup_introns_gencode_v44.tsv", sep="\t")
df_accepted["class"] = 1
df_accepted

df_rejected = pd.read_csv("data/og_rejected_introns", sep="\t")
df_rejected["class"] = 0
df_rejected

df_merged = pd.concat([df_accepted, df_rejected])
df_merged = df_merged.drop(columns="Duplicate_Count")
df_merged

# Self-simulated false introns

The set of simulated false introns is generated using:
1. Simulate false read using pbsim3
a. ONT single pass `pbsim --strategy templ --method qshmm --qshmm ~/Desktop/pbsim3-3.0.0/data/QSHMM-ONT.model --template ~/Desktop/pbsim3-3.0.0/gencode.v43.pc_transcripts.fa --prefix pbsim_rna_long --difference-ratio 39:24:36`
b. ONT multi-pass `pbsim --strategy templ --method qshmm --qshmm ~/Desktop/pbsim3-3.0.0/data/QSHMM-ONT.model --template ~/Desktop/pbsim3-3.0.0/gencode.v43.pc_transcripts.fa --prefix pbsim_rna_long_mp_3 --difference-ratio 39:24:36 --pass-num 3`
c. PacBio - Iso-seq `pbsim --strategy templ --method qshmm --qshmm ~/Desktop/pbsim3-3.0.0/data/QSHMM-RSII.model --template /home/rabbit/Desktop/pbsim3-3.0.0/gencode.v43.pc_transcripts.fa --prefix pbsim_rna_long_mp_3 --difference-ratio 22:45:33 --accuracy-mean 99.9 -pass-num 3`

2. Read-alignment using Minimap2
a. `minimap2 -ax splice -t 8 -uf -k14 ~/Documents/Projects/ML_gene_annot/release-109-hg38/hg38.fa ~/Documents/Projects/ML_gene_annot/pbsim3/pbsim_rna_long.fastq > aln.sam`
b. `samtools fastq ~/Documents/Projects/ML_gene_annot/pbsim3/ONT_real/pbsim_rna_long_mp_3.sam > simulated_ont_3pass.fq`  then  `minimap2 -ax splice -t 8 -u f -k 14 ~/Documents/Projects/ML_gene_annot/release-109-hg38/hg38.fa ~/Documents/Projects/ML_gene_annot/minimap2/ONT_real/simulated_ont_3pass.fq > aln2.sam`
c. `samtools fastq ~/Documents/Projects/ML_gene_annot/pbsim3/PacBio/pbsim_pacbio_rna_long_mp_3.sam > simulated_pacbio_3pass.fq` then `minimap2 -ax splice:hq -uf ~/Documents/Projects/ML_gene_annot/release-109-hg38/hg38.fa ~/Documents/Projects/ML_gene_annot/minimap2/PacBio/simulated_pacbio_3pass.fq > aln_pacbio.sam`

3. Sorted and convert to BAM
samtools sort ~/Documents/Projects/ML_gene_annot/minimap2/alignment.sam > aln_sorted.bam

4. Indexing
samtools index aln_sorted.bam

5. Get introns.ipynb

df_simulated_introns1 = pd.read_csv("./data/false_introns_simulated/introns.bed", sep="\t", names=["chr", "start", "end", "features", "score", "strand"])
df_simulated_introns1 = df_simulated_introns1.loc[:, ["chr", "start", "end", "strand"]]
df_simulated_introns1["class"] = 0
df_simulated_introns1 = df_simulated_introns1.drop_duplicates(subset=["chr", "start", "end", "strand"])
df_simulated_introns1

df_simulated_introns2 = pd.read_csv("./data/false_introns_simulated/introns2.bed", sep="\t", names=["chr", "start", "end", "features", "score", "strand"])
df_simulated_introns2 = df_simulated_introns2.loc[:, ["chr", "start", "end", "strand"]]
df_simulated_introns2["class"] = 0
df_simulated_introns2 = df_simulated_introns2.drop_duplicates(subset=["chr", "start", "end", "strand"])
df_simulated_introns2

df_simulated_introns3 = pd.read_csv("./data/false_introns_simulated/introns_pacbio_1.bed", sep="\t", names=["chr", "start", "end", "features", "score", "strand"])
df_simulated_introns3 = df_simulated_introns3.loc[:, ["chr", "start", "end", "strand"]]
df_simulated_introns3["class"] = 0
df_simulated_introns3 = df_simulated_introns3.drop_duplicates(subset=["chr", "start", "end", "strand"])
df_simulated_introns3

df_simulated_introns4 = pd.read_csv("./data/false_introns_simulated/introns_pacbio_2.bed", sep="\t", names=["chr", "start", "end", "features", "score", "strand"])
df_simulated_introns4 = df_simulated_introns4.loc[:, ["chr", "start", "end", "strand"]]
df_simulated_introns4["class"] = 0
df_simulated_introns4 = df_simulated_introns4.drop_duplicates(subset=["chr", "start", "end", "strand"])
df_simulated_introns4

merged_df = pd.concat([df_merged, df_simulated_introns2, df_simulated_introns3, df_simulated_introns4]).drop_duplicates(subset=["chr", "start", "end", "strand"], keep='first')
sum(merged_df["class"] == 0)

assert(len(merged_df[merged_df.start == 15038]) == 1)
assert(merged_df[merged_df.start == 15038].iloc[0]["class"] == 1)

 Now we have 125,464 false introns!

merged_df.to_csv("data/2-merged_train_set", sep="\t", index=False)
merged_df
