==> 04_combining_positive_and_negative_train_set.ipynb <==
import pandas as pd

df_gencode = pd.read_csv("data/01_gencode_introns.tsv", sep="\t")
df_gencode["class"] = 1
df_gencode

df_manually_rejected_introns = pd.read_csv("data/02_manually_rejected_introns.tsv", sep="\t")
df_manually_rejected_introns["class"] = 0
df_manually_rejected_introns = df_manually_rejected_introns.drop_duplicates(subset=["chr", "start", "end", "strand"])
df_manually_rejected_introns

df_simulation_a_introns = pd.read_csv("./data/false_introns_simulated/simulation_a_introns.bed", sep="\t",
									names=["chr", "start", "end", "features", "score", "strand"])
df_simulation_a_introns = df_simulation_a_introns.loc[:, ["chr", "start", "end", "strand"]]
df_simulation_a_introns["class"] = 0
df_simulation_a_introns = df_simulation_a_introns.drop_duplicates(subset=["chr", "start", "end", "strand"])
df_simulation_a_introns

df_simulation_b_introns = pd.read_csv("./data/false_introns_simulated/simulation_b_introns.bed", sep="\t",
									names=["chr", "start", "end", "features", "score", "strand"])
df_simulation_b_introns = df_simulation_b_introns.loc[:, ["chr", "start", "end", "strand"]]
df_simulation_b_introns["class"] = 0
df_simulation_b_introns = df_simulation_b_introns.drop_duplicates(subset=["chr", "start", "end", "strand"])
df_simulation_b_introns

df_simulation_c_introns = pd.read_csv("./data/false_introns_simulated/simulation_c_introns.bed", sep="\t",
									names=["chr", "start", "end", "features", "score", "strand"])
df_simulation_c_introns = df_simulation_c_introns.loc[:, ["chr", "start", "end", "strand"]]
df_simulation_c_introns["class"] = 0
df_simulation_c_introns = df_simulation_c_introns.drop_duplicates(subset=["chr", "start", "end", "strand"])
df_simulation_c_introns

df_simulation_d_introns = pd.read_csv("./data/false_introns_simulated/simulation_d_introns.bed", sep="\t",
									names=["chr", "start", "end", "features", "score", "strand"])
df_simulation_d_introns = df_simulation_d_introns.loc[:, ["chr", "start", "end", "strand"]]
df_simulation_d_introns["class"] = 0
df_simulation_d_introns = df_simulation_d_introns.drop_duplicates(subset=["chr", "start", "end", "strand"])
df_simulation_d_introns

df_trainset_combined = pd.concat([df_gencode, df_manually_rejected_introns, df_simulation_a_introns, df_simulation_b_introns,
					   df_simulation_c_introns, df_simulation_d_introns]).drop_duplicates(subset=["chr", "start", "end", "strand"], keep='first')

df_trainset_combined

print("Total amount of generated negative introns:", sum(df_trainset_combined["class"] == 0))

assert (len(df_trainset_combined[df_trainset_combined.start == 15038]) == 1)
assert (df_trainset_combined[df_trainset_combined.start == 15038].iloc[0]["class"] == 1)

df_trainset_combined.to_csv("data/04_train_set.tsv", sep="\t", index=False)
df_trainset_combined

!sha512sum data/04_train_set.tsv
