==> 05_recount3_feature.ipynb <==
import pandas as pd
from pandarallel import pandarallel

df_trainset = pd.read_csv("data/04_train_set.tsv", sep="\t")
df_trainset

def matchrc(row):
    potential_matches = !tabix data/resources/recount3.bed.gz {row.chr}:{row.start}-{row.end} | cut -f 2,3,5

    potential_matches = [
        int(entry.split('\t')[2])
        for entry in potential_matches
        if row['start'] == int(entry.split('\t')[0]) and row['end'] == int(entry.split('\t')[1]) #["RC3-Score"]
    ]

    if potential_matches:
        return potential_matches[0]
    else:
        return 0

pandarallel.initialize(nb_workers=12)

%%time
df_test = df_trainset.copy().head(10)
df_test["RC3"] = df_test.apply(matchrc, axis=1)
df_test

%%time
df_trainset["recount3_score"] = df_trainset.parallel_apply(matchrc, axis=1)
df_trainset

# Optional backup:
df_trainset.to_csv("data/05_trainset_recount3_feature.tsv", sep="\t", index=False)

! sha512sum data/05_trainset_recount3_feature.tsv
