==> 06_repeat_region_feature.ipynb <==
import pandas as pd
from pandarallel import pandarallel

df_trainset = pd.read_csv("data/04_train_set.tsv", sep="\t")
df_trainset

# Generating repeat features
We will generate the start ss (Donor) and end ss (Acceptor) separately, in the final script it will be merged to reduce run-time


def match_repeat_features_start_ss(row):
    """
    We call the tabix command, which look at the bed file to see if the row (which are entries in the intron that we extracted) overlaps with the repeat
    regions detailed in the repeat_features.bed.gz, apart from the splice-site ({row.start+2}), we also look at the small-region that precedes (2 nt in the exon; row.start-2) the splice-site.
    """
    matched_rows = !tabix data/resources/repeat_features.bed.gz {row.chr}:{row.start-2}-{row.start+2} | cut -f4
    return list(set(matched_rows + []))

def match_repeat_features_end_ss(row):
    """
    We call the tabix command, which look at the bed file to see if the row (which are entries in the intron that we extracted) overlaps with the repeat
    regions detailed in the repeat_features.bed.gz, apart from the splice-site ({row.end-2}), we also look at the small-region that proceeds (2 nt in the exon; {row.end+2}) the splice-site.
    """
    matched_rows2 = !tabix data/resources/repeat_features.bed.gz {row.chr}:{row.end-2}-{row.end+2} | cut -f4
        # | cut -f4 | sort -u
    return list(set([] + matched_rows2))

pandarallel.initialize(nb_workers=12, progress_bar=False)

# %%time
# # Test cell
# df_draft = df_trainset.copy().head(20)
# df_draft["Repeat_overlap"] = df_draft.apply(match_repeat_features_start_ss, axis=1)
# df_draft

%%time
df_trainset["repeat_features_start_site"] = df_trainset.parallel_apply(match_repeat_features_start_ss, axis=1)
df_trainset

%%time
df_trainset["repeat_features_end_site"] = df_trainset.parallel_apply(match_repeat_features_end_ss, axis=1)
df_trainset

df_trainset.to_csv("data/06_trainset_repeat_feature.tsv", sep="\t", index=False)

!sha512sum data/06_trainset_repeat_feature.tsv
