==> 09_nearest_alternative_splice_site_distance_feature.ipynb <==
import pandas as pd
import numpy as np
from pandarallel import pandarallel
from tqdm import tqdm

df_gencode = pd.read_csv("data/01_gencode_introns.tsv", sep="\t")
df_gencode = df_gencode[~(df_gencode.end - df_gencode.start <= 4)]
df_gencode = df_gencode.set_index(['strand', 'start', 'end'])
df_gencode = df_gencode.sort_index()
df_gencode

chrom_uniques = list(df_gencode["chr"].unique())  # we get the list of unique chromosome names
dict_chr = {}
# Splitting the dataframe chromosome groups separately into key(chromosome)-value(rows) in a dictionary
for chrom in tqdm(chrom_uniques):
	dict_chr[chrom] = df_gencode[df_gencode.chr == chrom]

dict_chr

# Sample query: Find the nearest (and not exact) entry of chr1 with strand '+' and nearest to 13579
# query_chr = 'chr1'
# query_strand = '-'
# query_position = 14499

def find_nearest_match_start_ss(row, site):
    query_chr = row["chr"]
    query_strand = row["strand"]
    query_position = row[site]

    # Filter the DataFrame for the relevant entries
    filtered_df = dict_chr[query_chr].loc[query_strand].copy()

    # Calculate the distance to the specified position
    filtered_df['distance'] = np.abs(filtered_df.index.get_level_values(site) - query_position)

    # Exclude exact match from the DataFrame (if it exists)
    filtered_df = filtered_df[filtered_df['distance'] != 0]

    # Find the row with the minimum distance (i.e., the nearest entry after excluding exact matches)
    nearest_entry = filtered_df.loc[filtered_df['distance'].idxmin()]

    return nearest_entry.distance

sha512sum = !sha512sum data/08_trainset_with_seq.tsv
assert("159da72ef546de9f15bfc662d51b70987134035499c64674f4240279ed9d120d34b0312bf98f1cda8c4d43222df7175f818c6d53c9dde2487ea7cf747b51ab05" in sha512sum[0])
df_training_data = pd.read_csv("data/08_trainset_with_seq.tsv", sep="\t", usecols=["chr","start","end","strand"])
df_training_data

pandarallel.initialize(nb_workers=12)

# %%time
# df_test = df_training_data.copy().head(10000)
# df_test["start_site_nearest"] = df_test.parallel_apply(find_nearest_match_start_ss, args=("start",), axis=1)
# df_test

%%time
df_training_data["nearest_alt_start_ss_dist"] = df_training_data.parallel_apply(find_nearest_match_start_ss, args=("start",), axis=1)

%%time
df_training_data["nearest_alt_end_ss_dist"] = df_training_data.parallel_apply(find_nearest_match_start_ss, args=("end",), axis=1)

df_training_data

df_training_data.to_csv("data/09_trainset_nearest_alt_ss_dist_feature", sep="\t", index=False)

!sha512sum data/09_trainset_nearest_alt_ss_dist_feature
