==> 10_maxentscan_score_feature.ipynb <==
import pandas as pd
import subprocess
from pandarallel import pandarallel

sha512sum = !sha512sum data/08_trainset_with_seq.tsv
assert("159da72ef546de9f15bfc662d51b70987134035499c64674f4240279ed9d120d34b0312bf98f1cda8c4d43222df7175f818c6d53c9dde2487ea7cf747b51ab05" in sha512sum[0])

df_train_set = pd.read_csv("data/08_trainset_with_seq.tsv", sep="\t", usecols=["maxentscan_sequence"])
df_train_set

df_train_set["mes_start_ss_seq"] = df_train_set["maxentscan_sequence"].apply(lambda x: x[:9])
df_train_set

df_train_set["mes_end_ss_seq"] = df_train_set["maxentscan_sequence"].apply(lambda x: x[-23:])
df_train_set

df_train_set=df_train_set.drop(columns=["maxentscan_sequence"], axis=1)
df_train_set

pandarallel.initialize()

def maxentscan_5(row):
    """
    We call the tabix command, which look at the bed file to see if the row (which are entries in the intron that we extracted) overlaps with the repeat
    regions detailed in the repeat_features.bed.gz, apart from the splice-site ({row.end-2}), we also look at the small-region that proceeds (2 nt in the exon; {row.end+2}) the splice-site.
    """
    # score = !perl MaxEntScan/score5.pl {row.sequence[:9]} | cut -f2
    # # score = score[-1:1]
    # return score
    target_sequence = row
    if len(target_sequence) < 9:
	    return 0
    result = subprocess.run(f"perl MaxEntScan/score5.pl {target_sequence}", shell=True, stdout=subprocess.PIPE, text=True)
    try:
    	score = result.stdout.strip().split("\t")[1]
    except:
        print(result)
        return result
    return score

# %%time
# df_test = df_train_set.copy().head(10)
# df_test["test"] = df_test["maxentscan_start_ss"].parallel_apply(maxentscan_5)
# df_test

%%time
df_train_set["MaxEntScan_start_ss"] = df_train_set["mes_start_ss_seq"].parallel_apply(maxentscan_5)

df_train_set.loc[df_train_set["mes_start_ss_seq"].str.contains("N"), "MaxEntScan_start_ss"] = 0
df_train_set[df_train_set["mes_start_ss_seq"].str.contains("N")]

# 3-prime end

def maxentscan_3(row):
    """
    We call the tabix command, which look at the bed file to see if the row (which are entries in the intron that we extracted) overlaps with the repeat
    regions detailed in the repeat_features.bed.gz, apart from the splice-site ({row.end-2}), we also look at the small-region that proceeds (2 nt in the exon; {row.end+2}) the splice-site.
    """
    target_sequence = row
    if len(target_sequence) < 23:
	    return 0
    result = subprocess.run(f"perl MaxEntScan/score3.pl {target_sequence}", shell=True, stdout=subprocess.PIPE, text=True)
    try:
    	score = result.stdout.strip().split("\t")[1]
    except:
        return 0
    return score

%%time
df_train_set["MaxEntScan_end_ss"] = df_train_set["mes_end_ss_seq"].parallel_apply(maxentscan_3)

df_train_set.loc[df_train_set["mes_end_ss_seq"].str.contains("N"), "mes_end_ss_seq"] = 0
df_train_set[df_train_set["mes_end_ss_seq"].str.contains("N")]

# Output to .tsv

df_train_set.to_csv("data/10_trainset_MaxEntScan_feature.tsv", sep="\t", index=False)

!sha512sum data/10_trainset_MaxEntScan_feature.tsv
