==> 11_basic_sequence_features.ipynb <==
import pandas as pd

df_merged = pd.read_csv("data/08_trainset_with_seq.tsv", sep="\t")
df_merged

df_merged["sequence"] = df_merged.maxentscan_sequence.str[3:-3]

def calculate_gc_content(sequence):
    return (sequence.count("G") + sequence.count("C")) / len(sequence)

df_merged["GC_content"] = df_merged["sequence"].apply(calculate_gc_content)

def calculate_cpg_island(sequence):
    c_count = sequence.count("C")
    g_count = sequence.count("G")
    cg_count = sequence.count("CG")
    total_count = len(sequence)

    try:
        cpg_ratio = (cg_count * total_count) / (c_count * g_count)
        gc_content = (c_count + g_count) / total_count
    except ZeroDivisionError:
        # In case there are sequences with no "C" or "G" nucleotides
        cpg_ratio = 0
        gc_content = 0

    return cpg_ratio > 0.6 and gc_content > 0.5

df_merged["CpG_island"] = df_merged["sequence"].apply(calculate_cpg_island)

def calculate_intron_length(row):
    return row.end - row.start

df_merged["intron_length"] = df_merged.apply(calculate_intron_length, axis=1)
df_merged

df_merged.to_csv("data/11_trainset_basic_features.tsv", sep="\t", index=False)

!md5sum data/11_trainset_basic_features.tsv
