==> 12_conservation_scores_feature.ipynb <==
import pyBigWig
import pandas as pd

md5sum = !md5sum data/conservation_scores/hg38.phyloP100way.bw | cut -f1
assert("43858006bdf98145b6fd239490bd0478" in md5sum[0])

bw_phyloP = pyBigWig.open("data/conservation_scores/hg38.phyloP100way.bw")

df_trainset = pd.read_csv("data/04_train_set.tsv", sep="\t")

## phyloP score feature

def phyloP_score(row):
    stats_start = bw_phyloP.stats(row.chr, row.start, row.start+2, exact=True)[0]
    stats_start = 0 if stats_start is None else stats_start
    stats_end = bw_phyloP.stats(row.chr, row.end-2, row.end, exact=True)[0]
    stats_end = 0 if stats_end is None else stats_end
    return (stats_start + stats_end)/2

%%time
df_trainset["phyloP_score"] = df_trainset.apply(phyloP_score, axis=1)
df_trainset

## phastCons score feature

bw_phastCons = pyBigWig.open("data/conservation_scores/hg38.phastCons100way.bw")

def phastCons_score(row):
    stats_start = bw_phastCons.stats(row.chr, row.start, row.start+2, exact=True)[0]
    stats_start = 0 if stats_start is None else stats_start
    stats_end = bw_phastCons.stats(row.chr, row.end-2, row.end, exact=True)[0]
    stats_end = 0 if stats_end is None else stats_end
    return (stats_start + stats_end)/2

%%time
df_trainset["phastCons_score"] = df_trainset.apply(phastCons_score, axis=1)
df_trainset

df_trainset.to_csv("data/12_trainset_conservation_scores_feature.tsv", sep="\t", index=False)

!md5sum data/12_trainset_conservation_scores_feature.tsv
