==> 13_recount3_near_ss_with_better_score_feature.ipynb <==
import pandas as pd
from pandarallel import pandarallel

df = pd.read_csv("data/05_trainset_recount3_feature.tsv", sep="\t", usecols=["chr","start","end","strand","recount3_score"])
df

gap = 5

# Start splice site (5")

def better_rc3_match_start_ss(row):
    potential_matches = !tabix data/resources/recount3.bed.gz {row.chr}:{row.start-gap}-{row.start+gap} | cut -f 2,5

    potential_matches = [
        (int(entry.split('\t')[0]), int(entry.split('\t')[1]))
        for entry in potential_matches
        if row['start'] - 5 <= int(entry.split('\t')[0]) <= row['start'] + 5 and int(entry.split('\t')[1]) > row["recount3_score"]
    ]
    if potential_matches:
        return True
    return False

pandarallel.initialize(nb_workers=12)

%%time
df_test = df.copy().head(10000)
df_test["better_match_5"] = df_test.parallel_apply(better_rc3_match_start_ss, axis=1)
df_test

%%time
df["recount3_near_start_ss_with_better_score"] = df.parallel_apply(better_rc3_match_start_ss, axis=1)
df

# End splice site (3")

def better_rc3_match_end_ss(row):
    potential_matches = !tabix data/resources/recount3.bed.gz {row.chr}:{row.end-gap}-{row.end+gap} | cut -f 2,5


    potential_matches = [
        (int(entry.split('\t')[0]), int(entry.split('\t')[1]))
        for entry in potential_matches
        if row['end'] - 5 <= int(entry.split('\t')[0]) <= row['end'] + 5 and int(entry.split('\t')[1]) > row["recount3_score"]
    ]
    if potential_matches:
        return True
    return False

%%time
df["recount3_near_end_ss_with_better_score"] = df.parallel_apply(better_rc3_match_end_ss, axis=1)
df

# output
df.to_csv("data/13_recount3_near_ss_with_better_score_feature.tsv", sep="\t", index=False)
