==> 14_BPP_prediction_feature.ipynb <==
import pandas as pd
from Bio import SeqIO
from Bio.Seq import Seq
from Bio.SeqRecord import SeqRecord

df_trainset = pd.read_csv("data/08_trainset_with_seq.tsv", sep="\t")
df_trainset

df_trainset = df_trainset[~(df_trainset.maxentscan_sequence.str.len() <= 24)]
df_trainset

# Create an empty list to hold the SeqRecord objects
seq_records = []

# Iterate through each row in the DataFrame
for index, row in df_trainset.iterrows():
    # Get the chromosome, start, and end positions
    chr_name = row['chr']
    start = row['start']
    end = row['end']
    strand = row['strand']

    # Get the sequence and remove the first and last 3 nucleotides
    seq = row['maxentscan_sequence'][3:-3]

    # Create a SeqRecord object
    seq_record = SeqRecord(Seq(seq),
                           id=f"{chr_name};{start};{end};{strand}",
                           description="")

    # Append the SeqRecord object to seq_records (list)
    seq_records.append(seq_record)

# Export the list of SeqRecord objects as a FASTA file
with open("data/BPP/introns.fasta", "w") as output_handle:
    SeqIO.write(seq_records, output_handle, "fasta")

!md5sum data/BPP/introns.fasta

!python3 data/BPP/BP_PPT.py -b data/BPP/pwmBP_human.txt -p data/BPP/scPPT_human.txt -i data/BPP/introns.fasta > data/14_trainset_BPP_score_feature.tsv

df_BPP_len = len(pd.read_csv("data/14_trainset_BPP_score_feature.tsv", sep="\t"))

# Identify rows to skip: in this example, rows 2, 4, 6, 8, 10, etc. have extra headers
skip_rows = [i for i in range(2, df_BPP_len, 2)]

df_BPP = pd.read_csv("data/14_trainset_BPP_score_feature.tsv", sep="\t", skiprows=skip_rows)

id_col_split = df_BPP["#id"].str.replace(">", "").str.split(";")

df_BPP["chr"] = id_col_split.str[0]
df_BPP["start"] = id_col_split.str[1]
df_BPP["end"] = id_col_split.str[2]
df_BPP["strand"] = id_col_split.str[3]

df_BPP = df_BPP.drop(columns="#id")

df_BPP.to_csv("data/14_trainset_BPP_score_feature.tsv", sep="\t", index=False)

!md5sum data/14_trainset_BPP_score_feature.tsv
