==> 15_combining_all_features.ipynb <==
import pandas as pd
import ast

df_trainset = pd.read_csv("data/04_train_set.tsv", sep="\t")
df_trainset

df_features_dict = {"5": pd.read_csv("data/05_trainset_recount3_feature.tsv", sep="\t"),
                    "6": pd.read_csv("data/06_trainset_repeat_feature.tsv", sep="\t"),
                    "7": pd.read_csv("data/07_trainset_antisense_feature.tsv", sep="\t"),
                    "8": pd.read_csv("data/08_trainset_with_seq.tsv", sep="\t"),
                    "9": pd.read_csv("data/09_trainset_nearest_alt_ss_dist_feature", sep="\t"),
                    "10": pd.read_csv("data/10_trainset_MaxEntScan_feature.tsv", sep="\t"),
                    "11": pd.read_csv("data/11_trainset_splam_feature.tsv", sep="\t"),
                    "12": pd.read_csv("data/12_trainset_conservation_scores_feature.tsv", sep="\t"),
                    "13": pd.read_csv("data/13_recount3_near_ss_with_better_score_feature.tsv", sep="\t"),
                    "14": pd.read_csv("data/14_trainset_BPP_score_feature.tsv", sep="\t")}

df_merged = df_trainset.copy()
df_merged

# Loop over the dictionary and concatenating it to the merged df
for key in df_features_dict.keys():
    if "class" in str(df_features_dict[key].columns):
        df_features_dict[key] = df_features_dict[key].drop(columns="class", axis=1)
    if len(df_features_dict[key]) == len(df_trainset):
        if key != "10":
                df_merged = pd.concat([df_merged, df_features_dict[key].drop(columns=["chr", "start", "end", "strand"])], axis=1)
        else:
            df_merged = pd.concat([df_merged, df_features_dict[key]], axis=1)
    else:
        df_merged = df_merged.merge(df_features_dict[key], how="left", on=["chr", "start", "end", "strand"])

df_merged

# Populate the repeat_features into one-hots

df_merged['lst_repeat_features_start_site'] = df_merged['repeat_features_start_site'].apply(lambda x: ast.literal_eval(x) if len(x) > 2 else [])
df_merged['lst_repeat_features_end_site'] = df_merged['repeat_features_end_site'].apply(lambda x: ast.literal_eval(x) if len(x) > 2 else [])
df_merged

unique_repeat_features_values_start_site = set()
unique_repeat_features_values_end_site = set()
for idx, row in df_merged.iterrows():
    unique_repeat_features_values_start_site.update(row['lst_repeat_features_start_site'])
    unique_repeat_features_values_end_site.update(row['lst_repeat_features_end_site'])

unique_repeat_features_values_start_site

for repeat_feature in unique_repeat_features_values_start_site:
    df_merged["repeat_features_start_site_"+repeat_feature] = df_merged["repeat_features_start_site"].str.count("'"+repeat_feature+"'").astype(int)


for repeat_feature in unique_repeat_features_values_start_site:
    df_merged["repeat_features_end_site_"+repeat_feature] = df_merged["repeat_features_end_site"].str.count("'"+repeat_feature+"'").astype(int)

df_merged

df_merged = df_merged.drop(columns=["repeat_features_start_site", "repeat_features_end_site", "lst_repeat_features_start_site", "lst_repeat_features_end_site"])
df_merged

df_merged.to_csv("data/15_trainset_populated.tsv", sep="\t", index=False)

!md5sum data/15_trainset_populated.tsv
