==> 17_feature_selection.ipynb <==
from sklearn.feature_selection import RFECV
import pandas as pd
from xgboost import XGBClassifier
from sklearn import preprocessing

df_train_x = pd.read_csv("data/15_complete_trainset.tsv", sep="\t")
y = df_train_x.pop("class")

le = preprocessing.LabelEncoder()
le.fit(df_train_x["splice_site"])
df_train_x["splice_site"] = le.transform(df_train_x["splice_site"])

le = preprocessing.LabelEncoder()
le.fit(df_train_x["bps"])
df_train_x["bps"] = le.transform(df_train_x["bps"])

df_train_x = df_train_x.drop(["strand", "chr"], axis=1)

df_train_x = df_train_x.fillna(0)

df_train_x

%%time

estimator = XGBClassifier(n_jobs=20)
selector = RFECV(estimator, min_features_to_select=10, step=1, verbose=1, cv=5, scoring="matthews_corrcoef")
selector = selector.fit(df_train_x, y)

feature_lst = df_train_x.columns.tolist()

feature_rank_name_pairs = list(zip(feature_lst, selector.ranking_))
feature_rank_name_pairs

feature_rank_name_pairs = sorted(feature_rank_name_pairs, key=lambda x: x[1])
feature_rank_name_pairs

[feature[0] for feature in feature_rank_name_pairs]

Based on the result above and consulting the annotation team for domain knowledge, we also manually pruned away some features.

df_final = pd.read_csv("data/15_complete_trainset.tsv", sep="\t")[['recount3_score',
 'antisense_exon_start_ss',
 'antisense_exon_end_ss',
 'nearest_alt_start_ss_dist',
 'nearest_alt_end_ss_dist',
 'MaxEntScan_start_ss',
 'MaxEntScan_end_ss',
 'CpG_island',
 'intron_length',
 'phyloP_score',
 'phastCons_score',
 'recount3_near_start_ss_with_better_score',
 'recount3_near_end_ss_with_better_score',
 # 'sc', # This is a feature out of the BPP (Branch Point Prediction tool). This is removed during the interval validation process due it requires the intron to have certain length, shorter intron might be immediately be disadvantaged because it has no predicted score (i.e. 0 or other value that we filled it to)
 'repeat_features_start_site_Tandem repeats',
 'repeat_features_start_site_LTRs',
 'repeat_features_start_site_Satellite repeats',
 'repeat_features_start_site_Type II Transposons',
 'repeat_features_end_site_Tandem repeats',
 'repeat_features_end_site_LTRs',
 'repeat_features_end_site_Type I Transposons/SINE',
 'repeat_features_end_site_Satellite repeats',
 'repeat_features_end_site_Type I Transposons/LINE',
 'splice_site',
 'repeat_features_start_site_Type I Transposons/SINE', 'class']]

df_final.to_csv("data/17_selected_features_trainset.tsv", sep="\t", index=False)
