==> 18_hyperparameter_optimisation.ipynb <==
import pandas as pd
from xgboost import XGBClassifier
from sklearn import preprocessing
from sklearn.model_selection import GridSearchCV

X = pd.read_csv("data/17_selected_features_trainset.tsv", sep="\t")
y = X.pop("class")

le = preprocessing.LabelEncoder()
le.fit(X["splice_site"])
X["splice_site"] = le.transform(X["splice_site"])

X

X = X.fillna(0)

%%time
from bayes_opt import BayesianOptimization
parameter_grid = {
    'n_estimators': [50, 100, 150, 200],
    'learning_rate': [0.01, 0.1, 0.2, 0.3],
    'max_depth': [5, 6, 7],
    'min_child_weight': [1, 3, 5],
    'subsample': [0.8, 0.9, 1.0],
    'colsample_bytree': [0.8, 0.9, 1.0]
}

model = XGBClassifier(tree_method="gpu_hist")

grid_search = GridSearchCV(model, parameter_grid, scoring="matthews_corrcoef", cv=3, verbose=1, n_jobs=12)

grid_search.fit(X, y)

df_grid_search_result = pd.DataFrame(grid_search.cv_results_)
df_grid_search_result.to_csv("data/18_hyperparameter_opt_result.tsv", sep="\t")
df_grid_search_result

print("Best Parameters:\n", grid_search.best_params_)
