==> 20b_external_benchmarking_gencode.ipynb <==
# Abstract / Conclusion

|                      | Accuracy  |
|----------------------|-----------|
| ION (Standard mode)  | **0.913** |
| ION (Strict   mode)  | 0.149     |
| SPLAM<sup>1</sup>    | 0.762     |

We can see that overall ION (Standard) mode performs the best in all category except for the False Positive Rate (FPR), the Strict mode of ION performs the best in terms of FPR but sacrifices significantly the overall predictive performance of the model.

import pandas as pd
from sklearn.metrics import accuracy_score
import warnings
from pyfaidx import Fasta
from Bio.Seq import Seq

warnings.filterwarnings("ignore")

df = pd.read_csv("benchmarking/gencode_v46_new_introns.tsv", sep="\t")
df

df["u1"] = 0
df["class"] = 1
df["strand"] = df.pop("strand")
df

# This is downloaded in https://hgdownload.soe.ucsc.edu/goldenPath/hg38/bigZips/
genome = Fasta('data/human_ref_hg38_109/GRCh38.primary_assembly.genome.fa', sequence_always_upper=True)

# A simple lambda function for matching the chromosome, start and end
coords_to_dna = lambda start_c, end_c, chr_c: genome[chr_c][start_c - 1:end_c]

def find_sequence_maxentscan(row):
	chromosome = row["chr"]
	start = row["start"]
	stop = row["end"]
	strand = row["strand"]

	seq_find_sequence = str(coords_to_dna(int(start) + 1 - 3, int(stop) + 3, chromosome))

	if strand == "-":  #strand
		seq_find_sequence = Seq(seq_find_sequence)  # Encode the sequence into
		seq_find_sequence = seq_find_sequence.reverse_complement()

	seq_find_sequence = str(seq_find_sequence)
	middle_part = seq_find_sequence[3:-3].upper()

	return middle_part


df["sequence"] = df.apply(find_sequence_maxentscan, axis=1)
df

df["splice-site"] = df["sequence"].str[:2] + ":" + df["sequence"].str[-2:]
df

df = df[(df["splice-site"] == "GT:AG") | (df["splice-site"] == "GC:AG") | (df["splice-site"] == "AT:AC")]
df

df = df.drop(columns=["splice-site", "sequence"])
df.to_csv("benchmarking/gencode_v46_new_introns.bed", sep="\t", header=False, index=False)

# Benchmarking using Splam: a deep-learning-based splice site predictor that improves spliced alignments
bioRxiv. 2023 Jul 29;2023.07.27.550754. doi: 10.1101/2023.07.27.550754. Preprint

!splam score -G ./data/human_ref_hg38_109/GRCh38.primary_assembly.genome.fa -m ./data/splam/splam_script.pt -o splam_out benchmarking/gencode_v46_new_introns.bed

df_splam_pred = pd.read_csv("benchmarking/splam_out_gencode_v46/junction_score.bed", sep="\t", names=["chr","start","end",'rc3_score','class','strand','donor_score','acceptor_score','u']).drop(columns="u")
df_splam_pred

df_splam_pred["splam_pred"] = ((df_splam_pred['donor_score'] >= 0.5) & (df_splam_pred['acceptor_score'] >= 0.5)).astype(int)
df_splam_pred

df_splam_pred["splam_pred_avg"] = df_splam_pred[['donor_score', 'acceptor_score']].min(axis=1)
df_splam_pred

acc = accuracy_score(df_splam_pred["class"], df_splam_pred["splam_pred"])

print("Metrics for SPLAM")
print("Accuracy: ", acc)

# ION (High-Recall Mode)

df_ion1 = pd.read_csv("benchmarking/ION_out_gencode_v46/output_high_recall_mode.bed", sep="\t")
df_ion1["prediction_class"] = (df_ion1["prediction"] >= 0.5).astype(int)
df_ion1["class"] = df_ion1["u2"]
acc = accuracy_score(df_ion1["class"], df_ion1["prediction_class"])

print("Metrics for ION (standard mode)")
print("Accuracy: ", acc)

# ION (Standard Mode)

df_ion2 = pd.read_csv("benchmarking/ION_out_gencode_v46/output_standard_mode.bed", sep="\t")
df_ion2["prediction_class"] = (df_ion2["prediction"] >= 0.5).astype(int)
df_ion2["class"] = df_ion2["u2"]
acc = accuracy_score(df_ion2["class"], df_ion2["prediction_class"])

print("Metrics for ION (standard mode)")
print("Accuracy: ", acc)

# ION (High-Precision Mode)

df_ion3 = pd.read_csv("benchmarking/ION_out_gencode_v46/output_high_precision_mode.bed", sep="\t")
df_ion3["prediction_class"] = (df_ion3["prediction"] >= 0.5).astype(int)
df_ion3["class"] = df_ion3["u2"]
acc = accuracy_score(df_ion3["class"], df_ion3["prediction_class"])

print("Metrics for ION (standard mode)")
print("Accuracy: ", acc)
